refactor: split operator_commands_ooda.rs into sub-module directory
Split the 616-line flat file into src/operator_commands_ooda/ with:
- mod.rs (8 lines): module declarations and pub re-exports
- daemon.rs (117 lines): run_ooda_daemon public function
- persistence.rs (45 lines): persist_cycle_report, persist_cycle_to_memory
- tests/mod.rs (89 lines): shared test helpers
- tests/persistence_tests.rs (133 lines): persistence unit tests
- tests/report_tests.rs (238 lines): report/summary/config tests

All files ≤238 lines (target was ≤400). The single pub export
(run_ooda_daemon) is re-exported from mod.rs so external callers
(operator_cli/ooda.rs) are unaffected.

Follows the same pattern as src/identity/.
